Update (module): Update
This will correct the optional gossip no longer working for some users. In debugger it kept going to case zero even thought it was designated case 11, not certain why so we do some constexpr auto to ensure it works specifically.
Final quests Where Kings Walk Quest 13188 and Saurfang's Blessing Quest 13189 are now auto completed as the quest being awarded with the item thru the module working but there was some issue with the quest when being awarded thru the module (edge case) so we just reward it (autocompleted) to ensure that the skip dk module is now fully skipped