Update (module): Rewrite and Optimizations
This is a Majority rewrite 99% worth by @danjohnso
Very Little guidence was provided by me.
Quest complete and rewarded for all quests of the dk starter spot including the racial specific one A Welcome Surprise quest variation.
Talent Points no appropriately reflected without the need to relog.
SaveAllPlayers is changed up to Just SaveToDB for just the player to assist with dead locks.
Confirmation for Skip is now a pop-up message.
Items are no longer auto equipped. All items are now in the player's bag, it will be up to the player to decide what to equip and get rid of.

Co-Authored-By: Dan Johnson <852995+danjohnso@users.noreply.github.com>